Refactor out methods for separation of concerns and testability
